+++ src/bscPlugin/validation/BrsFileValidator.spec.ts
@@ -6,6 +6,7 @@ import { DiagnosticMessages } from '../../DiagnosticMessages';
 import { expectDiagnostics, expectZeroDiagnostics } from '../../testHelpers.spec';
 import { Program } from '../../Program';
 import { isClassStatement, isNamespaceStatement } from '../../astUtils/reflection';
+import util from '../../util';
 
 describe('BrsFileValidator', () => {
     let program: Program;
@@ -84,5 +85,238 @@ describe('BrsFileValidator', () => {
                 DiagnosticMessages.keywordMustBeDeclaredAtNamespaceLevel('namespace')
             ]);
         });
+    });
+
+    it('allows classes in correct locations', () => {
+        program.setFile('source/main.bs', `
+            class Alpha
+            end class
+            namespace Beta
+                class Charlie
+                end class
+                namespace Delta
+                    class Echo
+                    end class
+                end namespace
+            end namespace
+        `);
+        program.validate();
+        expectZeroDiagnostics(program);
+    });
+
+    it('flags classes in wrong locations', () => {
+        program.setFile('source/main.bs', `
+            function test()
+                class Alpha
+                end class
+                if true then
+                    class Beta
+                    end class
+                end if
+            end function
+        `);
+        program.validate();
+        expectDiagnostics(program, [{
+            ...DiagnosticMessages.keywordMustBeDeclaredAtNamespaceLevel('class'),
+            range: util.createRange(2, 16, 2, 27)
+        }, {
+            ...DiagnosticMessages.keywordMustBeDeclaredAtNamespaceLevel('class'),
+            range: util.createRange(5, 20, 5, 30)
+        }]);
+    });
+
+    it('allows enums in correct locations', () => {
+        program.setFile('source/main.bs', `
+            enum Alpha
+                value1
+            end enum
+            namespace Beta
+                enum Charlie
+                    value1
+                end enum
+                namespace Delta
+                    enum Echo
+                        value1
+                    end enum
+                end namespace
+            end namespace
+        `);
+        program.validate();
+        expectZeroDiagnostics(program);
+    });
+
+    it('flags enums in wrong locations', () => {
+        program.setFile('source/main.bs', `
+            function test()
+                enum Alpha
+                    value1
+                end enum
+                if true then
+                    enum Beta
+                        value1
+                    end enum
+                end if
+            end function
+        `);
+        program.validate();
+        expectDiagnostics(program, [{
+            ...DiagnosticMessages.keywordMustBeDeclaredAtNamespaceLevel('enum'),
+            range: util.createRange(2, 16, 2, 26)
+        }, {
+            ...DiagnosticMessages.keywordMustBeDeclaredAtNamespaceLevel('enum'),
+            range: util.createRange(6, 20, 6, 29)
+        }]);
+    });
+
+    it('allows functions in correct locations', () => {
+        program.setFile('source/main.bs', `
+            function Alpha()
+            end function
+            namespace Beta
+                function Charlie()
+                end function
+                namespace Delta
+                    function Echo()
+                    end function
+                end namespace
+            end namespace
+        `);
+        program.validate();
+        expectZeroDiagnostics(program);
+    });
+
+    it('flags functions in wrong locations', () => {
+        program.setFile('source/main.bs', `
+            function test()
+                function Alpha()
+                end function
+                if true then
+                    function Beta()
+                    end function
+                end if
+            end function
+        `);
+        program.validate();
+        expectDiagnostics(program, [{
+            ...DiagnosticMessages.keywordMustBeDeclaredAtNamespaceLevel('function'),
+            range: util.createRange(2, 16, 2, 30)
+        }, {
+            ...DiagnosticMessages.keywordMustBeDeclaredAtNamespaceLevel('function'),
+            range: util.createRange(5, 20, 5, 33)
+        }]);
+    });
+
+    it('allows namespaces in correct locations', () => {
+        program.setFile('source/main.bs', `
+            namespace Alpha
+            end namespace
+            namespace Beta
+                namespace Charlie
+                end namespace
+                namespace Delta
+                    namespace Echo
+                    end namespace
+                end namespace
+            end namespace
+        `);
+        program.validate();
+        expectZeroDiagnostics(program);
+    });
+
+    it('flags classes in wrong locations', () => {
+        program.setFile('source/main.bs', `
+            function test()
+                namespace Alpha
+                end namespace
+                if true then
+                    namespace Beta
+                    end namespace
+                end if
+            end function
+        `);
+        program.validate();
+        expectDiagnostics(program, [{
+            ...DiagnosticMessages.keywordMustBeDeclaredAtNamespaceLevel('namespace'),
+            range: util.createRange(2, 16, 2, 31)
+        }, {
+            ...DiagnosticMessages.keywordMustBeDeclaredAtNamespaceLevel('namespace'),
+            range: util.createRange(5, 20, 5, 34)
+        }]);
+    });
+
+    it('allows interfaces in correct locations', () => {
+        program.setFile('source/main.bs', `
+            interface Alpha
+                prop as string
+            end interface
+            namespace Beta
+                interface Charlie
+                    prop as string
+                end interface
+                namespace Delta
+                    interface Echo
+                        prop as string
+                    end interface
+                end namespace
+            end namespace
+        `);
+        program.validate();
+        expectZeroDiagnostics(program);
+    });
+
+    it('flags interfaces in wrong locations', () => {
+        program.setFile('source/main.bs', `
+            function test()
+                interface Alpha
+                    prop as string
+                end interface
+                if true then
+                    interface Beta
+                        prop as string
+                    end interface
+                end if
+            end function
+        `);
+        program.validate();
+        expectDiagnostics(program, [{
+            ...DiagnosticMessages.keywordMustBeDeclaredAtNamespaceLevel('interface'),
+            range: util.createRange(2, 16, 2, 31)
+        }, {
+            ...DiagnosticMessages.keywordMustBeDeclaredAtNamespaceLevel('interface'),
+            range: util.createRange(6, 20, 6, 34)
+        }]);
+    });
+
+    it('allows consts in correct locations', () => {
+        program.setFile('source/main.bs', `
+            const Alpha = 1
+            namespace Beta
+                const Charlie = 2
+                namespace Delta
+                    const Echo = 3
+                end namespace
+            end namespace
+        `);
+        program.validate();
+        expectZeroDiagnostics(program);
+    });
+
+    it('flags consts in wrong locations', () => {
+        program.setFile('source/main.bs', `
+            function test()
+                const Alpha = 1
+                if true then
+                    const Beta = 2
+                end if
+            end function
+        `);
+        program.validate();
+        expectDiagnostics(program, [{
+            ...DiagnosticMessages.keywordMustBeDeclaredAtNamespaceLevel('const'),
+            range: util.createRange(2, 16, 2, 27)
+        }, {
+            ...DiagnosticMessages.keywordMustBeDeclaredAtNamespaceLevel('const'),
+            range: util.createRange(4, 20, 4, 30)
+        }]);
     });
 });